+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. [#5596](https://github.com/influxdata/chronograf/pull/5596): Show rule name from tickscript.
 1. [#5597](https://github.com/influxdata/chronograf/pull/5597): Do not truncate dashboard name with lots of room.
 1. [#5598](https://github.com/influxdata/chronograf/pull/5598): Repair TICKscript editor scrolling.
+1. [#5600](https://github.com/influxdata/chronograf/pull/5600): Apply default timeouts in server connections.
 
 ### Features